Updated with changes from OmniGroup svn [r209919]
Configurations:
- Support for replacing shell script build phases for BuildSVNVersionInclude with the Tools/SVNVersion project

SVNVersion:
- Project that does the work of the old BuildSVNVersionInclude shell script build phases while working better with project dependencies

OmniFoundation:
- Fix parsing of Korean dates with formats like '2014. 4. 20.'
- Restored some date parsing code from OmniFocus 1.x for falling back to English parsing when localized parsing doesn't match
- Added OFTransientObjectsTracker for tracking objects that are created and destroyed w/in a block of code
- Fix some implicit narrowing cast warnings

OmniInspector:
- Ensure standalone inspector UI is loaded before trying to restore expansion state

OmniSoftwareUpdate:
- Add support for automatically checking for updates on the run after a crash
- Don't offer to mail support for errors we can't help with (disk full, etc).

OmniAppKit:
- OATrackingLoop now tracks the current event in an ivar and reports it as a read-only property.
- Fix documentation window width at 800
- Updated OAStackView to notice when one of its subviews changes size,
  and to apply a constraint to itself when it wants to set its own
  height.
- Lots of improvments to OAWebPageViewer
- Add some API which lets us disable script toolbar items.
- Add some helpers on NSView for establishing same-extent constraints between a view and a contained view, or making those constraints and collecting them in an array.
- Fix first-day-of-week support in OAPopupDatePicker
- Don't try to produce tinted variants of template images.

OWF:
- When requesting data from a website, send a single preferred language
  rather than a complete ordered list of every language installed on the
  user's system (which some websites can't handle).
- A bit of modernization

OmniUI/iPad:
- iOS 7 updates to OUIInspectorOptionWheel
- OUIImagePickerAssetsViewController updates in response to library changes
- Reduce the number of undo operations logged for font size changes
- Add support for customizing appearance via user-supplied custom plists.
- Add some caching of appearance values
- Fix signature of blocks passed to imp_implementationWithBlock()
- Added helper +[UIView(OUIExtensions) topLevelViewFromNibNamed:]

OmniUI/iPad-Document:
- Don't try to unique file names when we are saving for overwriting
- Fix a few other edge cases with switching document UTIs while saving

Localizations:
- Various updates to strings files.

+++ Frameworks/OmniFoundation/CoreFoundationExtensions/CFData-OFCompression.m
@@ -1,4 +1,4 @@
-// Copyright 1997-2012 Omni Development, Inc. All rights reserved.
+// Copyright 1997-2012, 2014 Omni Development, Inc. All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -331,7 +331,7 @@ static Boolean readNullTerminatedString(FILE *fp,
             return FALSE;
         }
         
-        chBuf[0] = ch;
+        chBuf[0] = (UInt8)ch;
         CFDataAppendBytes(buffer, chBuf, 1);
     } while (ch != 0);
